Initial commit
This commit initializes the BattleCatGachaHelper project with the
necessary files and structure. This includes the BattleCatLib library
and the BattleCatWinForm application.

Current work include:
Simple UI layout.
Calculation of the future gacha pulls.
Ability to mark the units for planning.
Read data of cat units from a yaml file. (Not published yet)